refactor: remove timefoering period wrapper
